Refactored defaultComponentBuilder into a priority list of individual builders to better support configuration by the end-user.
